# back-end/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  api.CreateEventRequest:
    properties:
      actor_id:
        type: integer
      batch_id:
        type: integer
      event_type:
        type: string
      location:
        type: string
      metadata:
        additionalProperties: true
        type: object
    type: object
...
  api.ErrorResponse:
    properties:
      error:
        type: string
      error_detail:
        type: string
      error_type:
        type: string
      message:
        type: string
      method:
        type: string
      path:
        type: string
      request_id:
        type: string
      status_code:
        type: integer
      success:
        type: boolean
      timestamp:
        type: string
    type: object
...
  api.SuccessResponse:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
...
  api.UpdateEventRequest:
    properties:
      actor_id:
        type: integer
      event_type:
        type: string
      location:
        type: string
      metadata:
        additionalProperties: true
        type: object
    type: object
...
  models.Event:
    properties:
      actor:
        type: object
      actor_id:
        description: Refers to User.ID
        type: integer
      batch_id:
        description: Refers to Batch.ID
        type: integer
      blockchain_records:
        description: Related blockchain records
        items:
          type: object
        type: array
      event_type:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      location:
        type: string
      metadata:
        items:
          type: integer
        type: array
      timestamp:
        type: string
      updated_at:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@vietchain.com
    name: API Support
  description: Traceability system for shrimp larvae using blockchain technology
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: TracePost-larvaeChain API
  version: "1.0"
paths:
...
  /blockchain/event/{eventId}:
    get:
      consumes:
      - application/json
      description: Retrieve event data directly from the blockchain
      parameters:
      - description: Event ID
        in: path
        name: eventId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Get event from blockchain
      tags:
      - blockchain
...
  /events:
    get:
      consumes:
      - application/json
      description: Retrieve all event records with optional filtering
      parameters:
      - description: Filter by batch ID
        in: query
        name: batch_id
        type: integer
      - description: Filter by event type
        in: query
        name: event_type
        type: string
      - description: Filter by actor ID
        in: query
        name: actor_id
        type: integer
      - description: 'Limit number of results (default: 50)'
        in: query
        name: limit
        type: integer
      - description: 'Offset for pagination (default: 0)'
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api.SuccessResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Event'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Get all events
      tags:
      - events
    post:
      consumes:
      - application/json
      description: Create a new event for a shrimp larvae batch
      parameters:
      - description: Event creation details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.CreateEventRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/api.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.Event'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Create a new event
      tags:
      - events
  /events/{id}:
    delete:
      consumes:
      - application/json
      description: Soft delete an event (sets is_active to false)
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Delete event
      tags:
      - events
    get:
      consumes:
      - application/json
      description: Retrieve a specific event by its ID with related data
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.Event'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Get event by ID
      tags:
      - events
    put:
      consumes:
      - application/json
      description: Update an existing event's information
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: integer
      - description: Event update details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.UpdateEventRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.Event'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Update event
      tags:
      - events
...
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"